chapter4/server_primitives/host_copy/pyproject.toml
[project]
name = "host"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents==1.20.0",
    "streamlit==1.52.2",
]
